google_api.py
import datetime
import os.path

from google.auth.transport.requests import Request
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow
from googleapiclient.discovery import build
from googleapiclient.errors import HttpError

# If modifying these scopes, delete the file token.json.
SCOPES = ["https://www.googleapis.com/auth/calendar"]


def create_service(cliente_secret_file, api_name, api_version, *scopes, prefix=''):
  """Shows basic usage of the Google Calendar API.
  Prints the start and name of the next 10 events on the user's calendar.
  """
  CLIENT_SECRET_FILE = cliente_secret_file
  API_SERVICE_NAME = api_name
  API_VERSION = api_version
  SCOPES = [ scope for scope in scopes[0] ]

  cred = None

  working_dir = os.getcwd()
  token_dir = 'token files'
  token_file = f'token_{API_SERVICE_NAME}_{API_VERSION}-{prefix}.json'

  ### Check if token dir exists first, if not, create the folder
  if not os.path.exists(os.path.join(working_dir, token_dir)):
      os.mkdir(os.path.join(working_dir, token_dir))

  if os.path.exists(os.path.join(working_dir, token_dir, token_file)):
      cred = Credentials.from_authorized_user_file(os.path.join(working_dir, token_dir, token_file), SCOPES)

  if not cred or not cred.valid:
      if cred and cred.expired and cred.refresh_token:
          cred.refresh(Request())
      else:
          flow = InstalledAppFlow.from_client_secrets_file(CLIENT_SECRET_FILE, SCOPES)
          cred = flow.run_local_server(port=0)

      with open(os.path.join(working_dir, token_dir, token_file), 'w') as token:
          token.write(cred.to_json())

  try:
      service = build(API_SERVICE_NAME, API_VERSION, credentials=cred, static_discovery=False)
      print(API_SERVICE_NAME, API_VERSION, 'service created successfully')
      return service
  except Exception as e:
      print(e)
      print(f'Failed to create service instance for {API_SERVICE_NAME}')
      os.remove(os.path.join(working_dir, token_dir, token_file))
      return None